Reduce number of warnings for 64-bit targets.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@35272 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/config/dir_list.cpp
@@ -46,7 +46,7 @@ void ParsedPath::parse(const PathName& path)
 	} while (oldpath.length() > 0);
 }
 
-PathName ParsedPath::subPath(int n) const 
+PathName ParsedPath::subPath(size_t n) const 
 {
 	PathName rc = (*this)[0];
 	if (PathUtils::isRelative(rc + PathUtils::dir_sep))
@@ -153,7 +153,7 @@ void DirectoryList::initialize(bool simple_mode)
 		}
 	}
 
-	unsigned int last = 0;
+	size_t last = 0;
 	PathName root = Config::getRootDirectory();
 	size_t i;
 	for (i = 0; i < val.length(); i++) {
